requirements.txt
appdirs
asgiref
certifi
cffi
chardet
cryptography
defusedxml
distlib
Django
django-debug-toolbar
django-debug-toolbar-template-profiler
django-extensions
filelock
idna
MainApp
oauthlib
Pillow
pycparser
PyJWT
python3-openid
pytz
requests
requests-oauthlib
six
social-auth-app-django
social-auth-core
sqlparse
urllib3
virtualenv
psycopg2
